mgr/dashboard: Drop iSCSI gateway name parameter
iSCSI gateway name will now be obtained from the
ceph-iscsi `/api/sysinfo/hostname` endpoint
